+++ tests/testthat/test-add-variable-definition.R
@@ -59,7 +59,8 @@ if (run.integration.tests) {
             })
 
             test_that("Can insert VarDef with no values", {
-                expect_warning(ds$newvar3 <- VarDef(name="Empty", type="numeric"), "Adding variable with no rows of data")
+                expect_warning(ds$newvar3 <- VarDef(name="Empty", type="numeric"),
+                    "Adding variable with no rows of data")
                 expect_identical(as.vector(ds$newvar3), rep(NA_real_, 20L))
             })
 
@@ -114,7 +115,7 @@ if (run.integration.tests) {
                     expect_error(addVariables(ds,
                         VarDef(ds$v3 + 4, name="v3plus4"),
                         VarDef(ds$v3 + 5, name="v3plus5")
-                    ), "Error : The following variable definition\\(s\\) errored on upload: 1, 2")
+                    ), "The following variable definition\\(s\\) errored on upload: 1, 2")
                 })
                 ## Confirm that refresh(ds) is unchanged
                 ds <- refresh(ds)