+++ NEWS.md
@@ -1,4 +1,5 @@
 ### crunch 1.9.3 (under development)
+* Update tests to pass with forthcoming release of `testthat`.
 
 ### crunch 1.9.2
 * `forkDataset` to make a fork (copy) of a dataset; `mergeFork` to merge changes from a fork back to its parent (or vice versa)